Avoid hidden path prefix for project configuration file
Windows and UNIX based systems have different hidden path prefixes (`_`
vs `.`) so this means the user has to duplicate the configuration file
to include both prefixes if his project wants to support multiple
platforms correctly.

In order to not force a specific hidden prefix (for consistency), we
make the path to the project configuration file non hidden.